Frontend: Refactor NewRace formFields, validationSchema away from component
